+++ smart-contracts/contracts/Unlock.sol
@@ -1,4 +1,4 @@
-pragma solidity 0.5.7;
+pragma solidity 0.5.8;
 
 /**
  * @title The Unlock contract
@@ -196,7 +196,7 @@ contract Unlock is
   ) external pure
     returns (uint16)
   {
-    return 3;
+    return 4;
   }
 
   // function to read the globalTokenURI field.