+++ tickets/src/middlewares/web3Middleware.js
@@ -1,6 +1,6 @@
 /* eslint promise/prefer-await-to-then: 0 */
 
-import UnlockJs from '@unlock-protocol/unlock-js'
+import { Web3Service } from '@unlock-protocol/unlock-js'
 import { LOCATION_CHANGE } from 'connected-react-router'
 
 import { startLoading, doneLoading } from '../actions/loading'
@@ -23,8 +23,6 @@ import {
   signedAddressMismatch,
 } from '../actions/ticket'
 import UnlockEventRSVP from '../structured_data/unlockEventRSVP'
-
-const { Web3Service } = UnlockJs
 
 // This middleware listen to redux events and invokes the web3Service API.
 // It also listen to events from web3Service and dispatches corresponding actions
@@ -165,13 +163,13 @@ const web3Middleware = config => {
             eventAddress: eventAddress,
           })
 
-          web3Service.recoverAccountFromSignedData(
-            JSON.stringify(data),
-            signedAddress,
-            (error, account) => {
+          web3Service
+            .recoverAccountFromSignedData(JSON.stringify(data), signedAddress)
+            .then(account => {
               const normalizedAccount = account.toString().toLowerCase()
               const normalizedPublicKey = publicKey.toString().toLowerCase()
-              if (error || normalizedAccount !== normalizedPublicKey) {
+
+              if (normalizedAccount !== normalizedPublicKey) {
                 dispatch(
                   signedAddressMismatch(normalizedPublicKey, signedAddress)
                 )
@@ -184,8 +182,7 @@ const web3Middleware = config => {
                   )
                 )
               }
-            }
-          )
+            })
         }
 
         const keyId = `${lockAddress}-${accountAddress}`